Refactor images and update styles for header, footer, and pages; replace GIFs with PNGs for better performance and clarity.
